requirements.txt
python-telegram-bot==20.7
googletrans==4.0.0rc1
python-docx==1.1.0
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.31.0

